refactor: update button component props to use separate files for primary and secondary buttons
